requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
cmarkgfm==2022.10.27
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.54.0
django-appconf==1.0.5
django-debug-toolbar==4.1.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-multiselectfield==0.1.12
django-storages==1.13.2
django-taggit==4.0.0
django-tuieditor==0.2.0
executing==1.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.13.2
jedi==0.18.2
jmespath==1.0.1
Markdown==3.4.3
matplotlib-inline==0.1.6
msgpack==1.0.5
mysqlclient==2.1.1
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
urllib3==1.26.16
wcwidth==0.2.6
zope.interface==6.0
